feat : 플레이어 마나 , 피해 데미지 ui 추가
플레이어가 피해 입는 수치와 마나를 알기 위해 ui구현

+++ Assets/Scripts/Player/Player.cs
@@ -5,7 +5,7 @@ using UnityEngine.Events;
 
 public abstract class Player : MonoBehaviour,IDamageable
 {
-    public event UnityAction<float> onChangeHp;
+    public event UnityAction<int> onChangeHp;
